+++ libs/common/src/lib/upload/upload-qna/upload-qna.component.ts
@@ -4,7 +4,7 @@ import { StandaloneService } from '../../services';
 import { UploadService } from '../upload.service';
 import { STFTrackingService } from '@flaps/core';
 import { FormControl, Validators } from '@angular/forms';
-import { SisToastService } from '@nuclia/sistema';
+import { TextFormat } from '@nuclia/core';
 
 @Component({
   selector: 'nuclia-upload-qna',
@@ -21,6 +21,7 @@ export class UploadQnaComponent {
     nonNullable: true,
     validators: [Validators.required],
   });
+  qnaFormat = new FormControl<TextFormat>('PLAIN', { nonNullable: true });
   qna: string[][] = [];
 
   constructor(
@@ -29,7 +30,6 @@ export class UploadQnaComponent {
     private standaloneService: StandaloneService,
     private cdr: ChangeDetectorRef,
     private tracking: STFTrackingService,
-    private toaster: SisToastService,
   ) {}
 
   close(): void {
@@ -45,13 +45,15 @@ export class UploadQnaComponent {
     if (this.resourceTitle.valid && this.qna.length > 0) {
       this.tracking.logEvent('upload_q_and_a_from_csv');
       this.isUploading = true;
-      this.uploadService.uploadQnaResource(this.resourceTitle.getRawValue(), this.qna).subscribe({
-        next: () => this.modal.close(),
-        error: () => {
-          this.isUploading = false;
-          this.cdr.markForCheck();
-        },
-      });
+      this.uploadService
+        .uploadQnaResource(this.resourceTitle.getRawValue(), this.qna, this.qnaFormat.getRawValue())
+        .subscribe({
+          next: () => this.modal.close(),
+          error: () => {
+            this.isUploading = false;
+            this.cdr.markForCheck();
+          },
+        });
     }
   }
 }