+++ libs/search-widget/src/tiles/TextTile.svelte
@@ -2,8 +2,8 @@
   import type { FileField, Search, TextField } from '@nuclia/core';
   import type { WidgetParagraph } from '../core/models';
   import { PreviewKind } from '../core/models';
-  import TextViewer from './viewers/TextViewer.svelte';
-  import MarkdownViewer from './viewers/MarkdownViewer.svelte';
+  import ExtractedTextViewer from './viewers/ExtractedTextViewer.svelte';
+  import TextContentViewer from './viewers/TextContentViewer.svelte';
   import { getCDN } from '../core/utils';
   import DocumentTile from './base-tile/DocumentTile.svelte';
   import { fieldData } from '../core/stores/viewer.store';
@@ -23,6 +23,26 @@
     ),
   );
 
+  const isHtml = fieldData.pipe(
+    map(
+      (data) =>
+        (data?.value as TextField)?.format === 'HTML' ||
+        (data?.value as FileField).file?.content_type === 'text/html' ||
+        ((data?.value as FileField).file?.content_type === 'application/octet-stream' &&
+          (data?.value as FileField).file?.filename?.endsWith('.html')),
+    ),
+  );
+
+  const isRst = fieldData.pipe(
+    map(
+      (data) =>
+        (data?.value as TextField)?.format === 'RST' ||
+        (data?.value as FileField).file?.content_type === 'text/x-rst' ||
+        ((data?.value as FileField).file?.content_type === 'application/octet-stream' &&
+          (data?.value as FileField).file?.filename?.endsWith('.rst')),
+    ),
+  );
+
   function openParagraph(paragraph: WidgetParagraph) {
     selectedParagraph = paragraph;
   }
@@ -33,9 +53,13 @@
   fallbackThumbnail={`${getCDN()}icons/text/plain.svg`}
   {result}
   on:selectParagraph={(event) => openParagraph(event.detail)}>
-  {#if $isMarkdown}
-    <MarkdownViewer {selectedParagraph} />
+  {#if $isMarkdown || $isHtml}
+    <TextContentViewer
+      {selectedParagraph}
+      isHtml={$isHtml} />
   {:else}
-    <TextViewer {selectedParagraph} />
+    <ExtractedTextViewer
+      {selectedParagraph}
+      isRst={$isRst} />
   {/if}
 </DocumentTile>