+++ libs/search-widget/src/tiles/viewers/renderer/PlainTextRenderer.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  export let text = '';
+
+  $: formattedText = text.trim().replace(/\n/g, '<br>');
+</script>
+
+{@html formattedText}